% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LJCTrasferEntropy.R
\name{LJCTransferEntropy}
\alias{LJCTransferEntropy}
\title{Improved Transfer Entropy Analysis with Rolling Window}
\usage{
LJCTransferEntropy(
  x,
  y,
  Time,
  lx = 1,
  ly = 1,
  q = 0.1,
  entropy = "Shannon",
  shuffles = 100,
  type = "quantiles",
  quantiles = c(5, 95),
  bins = NULL,
  limits = NULL,
  nboot = 300,
  burn = 50,
  quiet = NULL,
  seed = NULL,
  na.rm = TRUE,
  eff_entropy = TRUE,
  cr = "AIC",
  start_window_size = 10,
  rolling_step_size = 1
)
}
\arguments{
\item{x}{Time series data for variable x.}

\item{y}{Time series data for variable y.}

\item{lx}{Embedding dimension for x.}

\item{ly}{Embedding dimension for y.}

\item{q}{Quantile for Transfer Entropy.}

\item{entropy}{Type of entropy to use.}

\item{shuffles}{Number of shuffles for significance testing.}

\item{type}{Type of binning.}

\item{quantiles}{Quantiles for binning.}

\item{bins}{Number of bins.}

\item{limits}{Custom limits for binning.}

\item{nboot}{Number of bootstrap iterations.}

\item{burn}{Burn-in period for bootstrap.}

\item{quiet}{Suppress output.}

\item{seed}{Random seed for reproducibility.}

\item{na.rm}{Remove missing values.}

\item{eff_entropy}{Use efficient entropy calculation.}

\item{cr}{Criterion for stationarity test.}
}
\value{
A data frame containing Transfer Entropy results for each window.
}
\description{
Improved Transfer Entropy Analysis with Rolling Window
}
